requirements.txt
pyyaml==6.0.1
requests==2.31.0
rich==13.7.1
questionary==2.0.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
aiohttp==3.9.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
